+++ post-expirator.php
@@ -5,12 +5,12 @@
  * Plugin URI: http://wordpress.org/extend/plugins/post-expirator/
  * Description: PublishPress Future allows you to schedule automatic changes to posts, pages and other content types.
  * Author: PublishPress
- * Version: 4.1.0-beta.3
+ * Version: 4.1.1
  * Author URI: http://publishpress.com
  * Text Domain: post-expirator
  * Domain Path: /languages
- * Requires at least: 6.1
- * Requires PHP: 7.2.5
+ * Requires at least: 6.7
+ * Requires PHP: 7.4
  */
 
 namespace PublishPress\Future;
@@ -26,8 +26,8 @@ defined('ABSPATH') or die('Direct access not allowed.');
 
 global $wp_version;
 
-$min_php_version = '7.2.5';
-$min_wp_version  = '6.1';
+$min_php_version = '7.4';
+$min_wp_version  = '6.7';
 
 // If the PHP or WP version is not compatible, terminate the plugin execution.
 $invalid_php_version = version_compare(phpversion(), $min_php_version, '<');
@@ -51,7 +51,7 @@ if (! defined('PUBLISHPRESS_FUTURE_LOADED')) {
         }
 
         if (! defined('PUBLISHPRESS_FUTURE_VERSION')) {
-            define('PUBLISHPRESS_FUTURE_VERSION', '4.1.0-beta.3');
+            define('PUBLISHPRESS_FUTURE_VERSION', '4.1.1');
         }
 
         if (! defined('PUBLISHPRESS_FUTURE_PLUGIN_FILE')) {