+++ src/Modules/Expirator/Module.php
@@ -17,6 +17,7 @@ use PublishPress\Future\Framework\WordPress\Facade\RequestFacade;
 use PublishPress\Future\Framework\Database\Interfaces\DBTableSchemaInterface;
 use PublishPress\Future\Framework\Logger\LoggerInterface;
 use PublishPress\Future\Framework\System\DateTimeHandlerInterface;
+use PublishPress\Future\Modules\Expirator\Models\PostTypeDefaultDataModelFactory;
 use PublishPress\Future\Modules\Settings\SettingsFacade;
 
 defined('ABSPATH') or die('Direct access not allowed.');
@@ -103,6 +104,16 @@ class Module implements ModuleInterface
      */
     private $dateTimeHandler;
 
+    /**
+     * @var PostTypeDefaultDataModelFactory
+     */
+    private $defaultDataModelFactory;
+
+    /**
+     * @var \Closure
+     */
+    private $taxonomiesModelFactory;
+
     public function __construct(
         \PublishPress\Future\Core\HookableInterface $hooks,
         SiteFacade $site,
@@ -118,7 +129,9 @@ class Module implements ModuleInterface
         DBTableSchemaInterface $actionArgsSchema,
         SettingsFacade $settingsFacade,
         LoggerInterface $logger,
-        DateTimeHandlerInterface $dateTimeHandler
+        DateTimeHandlerInterface $dateTimeHandler,
+        PostTypeDefaultDataModelFactory $defaultDataModelFactory,
+        \Closure $taxonomiesModelFactory
     ) {
         $this->hooks = $hooks;
         $this->site = $site;
@@ -135,6 +148,8 @@ class Module implements ModuleInterface
         $this->settingsFacade = $settingsFacade;
         $this->logger = $logger;
         $this->dateTimeHandler = $dateTimeHandler;
+        $this->defaultDataModelFactory = $defaultDataModelFactory;
+        $this->taxonomiesModelFactory = $taxonomiesModelFactory;
 
         $this->controllers['expiration'] = $this->factoryExpirationController();
         $this->controllers['quick_edit'] = $this->factoryQuickEditController();
@@ -166,7 +181,8 @@ class Module implements ModuleInterface
         return new Controllers\ExpirationController(
             $this->hooks,
             $this->scheduler,
-            $this->expirablePostModelFactory
+            $this->expirablePostModelFactory,
+            $this->logger
         );
     }
 
@@ -252,7 +268,8 @@ class Module implements ModuleInterface
             $this->expirablePostModelFactory,
             $this->currentUserModelFactory,
             $this->logger,
-            $this->dateTimeHandler
+            $this->dateTimeHandler,
+            $this->taxonomiesModelFactory
         );
     }